backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.20.0
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
PySocks==1.7.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
rpds-py==0.29.0
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
zope.event==6.1
zope.interface==8.1.1
